+++ package.json
@@ -39,8 +39,8 @@
     "webpack-dev-server": "^3.11.2"
   },
   "dependencies": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^17.0.0",
+    "react-dom": "^17.0.0",
     "react-redux": "^7.2.6",
     "redux": "^4.1.2"
   }
