+++ src/InputContainer.jsx
@@ -1,0 +1,32 @@
+import { useDispatch, useSelector } from 'react-redux';
+
+import Input from './Input';
+
+import {
+  updateTaskTitle,
+  addTask,
+} from './actions';
+
+function handleChangeTitle(event) {
+  useDispatch(updateTaskTitle(event.target.value));
+}
+
+function handleClickAddTask() {
+  useDispatch(addTask());
+}
+
+export default function InputContainer() {
+  const { taskTitle } = useSelector((state) => ({
+    taskTitle: state.taskTitle,
+  }));
+
+  return (
+    <div>
+      <Input
+        value={taskTitle}
+        onChange={handleChangeTitle}
+        onClick={handleClickAddTask}
+      />
+    </div>
+  );
+}
